requirements.txt
asgiref==3.8.1
attrs==23.2.0
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
Django==5.0.4
django-storages==1.14.2
djangorestframework==3.15.1
drf-spectacular==0.27.2
google-api-core==2.18.0
google-api-python-client==2.125.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-storage==2.16.0
google-crc32c==1.5.0
google-resumable-media==2.7.0
googleapis-common-protos==1.63.0
grpcio==1.62.1
grpcio-status==1.62.1
httplib2==0.22.0
idna==3.7
inflection==0.5.1
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
lxml==5.2.1
mysql-connector-python==8.3.0
pillow==10.3.0
proto-plus==1.23.0
protobuf==4.25.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
pyparsing==3.1.2
PyPDF2==3.0.1
python-docx==1.1.0
PyYAML==6.0.1
referencing==0.34.0
requests==2.31.0
rpds-py==0.18.0
rsa==4.9
sqlparse==0.4.4
typing_extensions==4.11.0
tzdata==2024.1
uploader==0.5.1
uritemplate==4.1.1
urllib3==2.2.1
